requirements.txt
autopep8
beautifulsoup4
chromadb
duckduckgo-search
google-generativeai
psutil
pytest
PyAudio
python-dotenv
gTTS
pydub
pypdf
SpeechRecognition
PyYAML
requests
fastapi
uvicorn
python-multipart
huggingface-hub>=0.10.0